+++ db/sql/verify/buildings_bulk_load/add_new_pk_column_to_removed.sql
@@ -1,0 +1,7 @@
+-- Verify nz-buildings:buildings_bulk_load/add_new_pk_column_to_removed on pg
+
+BEGIN;
+
+-- XXX Add verifications here.
+
+ROLLBACK;