+++ db/tests/testdata/db/buildings_bulk_load.sql
@@ -382,13 +382,13 @@ INSERT INTO buildings_bulk_load.related (related_id, related_group_id, bulk_load
 
 -- buildings_bulk_load.removed
 
-INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id) VALUES (1004, 1);
-INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id) VALUES (1006, 1);
-INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id) VALUES (1034, 1);
-INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id) VALUES (1042, 1);
-INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id) VALUES (1058, 1);
-INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id) VALUES (1064, 1);
-INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id) VALUES (1081, 1);
+INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id, supplied_dataset_id) VALUES (1004, 1, 1);
+INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id, supplied_dataset_id) VALUES (1006, 1, 1);
+INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id, supplied_dataset_id) VALUES (1034, 1, 2);
+INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id, supplied_dataset_id) VALUES (1042, 1, 4);
+INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id, supplied_dataset_id) VALUES (1058, 1, 4);
+INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id, supplied_dataset_id) VALUES (1064, 1, 4);
+INSERT INTO buildings_bulk_load.removed (building_outline_id, qa_status_id, supplied_dataset_id) VALUES (1081, 1, 4);
 
 -- buildings_bulk_load.supplied_outlines